olive-garden-res-ui-app/requirement.txt
streamlit>=1.30.0
boto3
pandas
openpyxl
Pillow